+++ netappfiles-go-sdk-sample/internal/sdkutils/sdkutils.go
@@ -15,12 +15,12 @@ import (
 	"strings"
 	"time"
 
-	"github.com/Azure-Samples/netappfiles-go-sdk-sample/internal/iam"
-	"github.com/Azure-Samples/netappfiles-go-sdk-sample/internal/uri"
-	"github.com/Azure-Samples/netappfiles-go-sdk-sample/internal/utils"
+	"github.com/Azure-Samples/netappfiles-go-sdk-sample/netappfiles-go-sdk-sample/internal/iam"
+	"github.com/Azure-Samples/netappfiles-go-sdk-sample/netappfiles-go-sdk-sample/internal/uri"
+	"github.com/Azure-Samples/netappfiles-go-sdk-sample/netappfiles-go-sdk-sample/internal/utils"
 
-	"github.com/Azure/azure-sdk-for-go/services/netapp/mgmt/2019-10-01/netapp"
-	"github.com/Azure/azure-sdk-for-go/services/resources/mgmt/2019-05-01/resources"
+	"github.com/Azure/azure-sdk-for-go/services/netapp/mgmt/2020-06-01/netapp"
+	"github.com/Azure/azure-sdk-for-go/services/resources/mgmt/2019-10-01/resources"
 	"github.com/Azure/go-autorest/autorest/to"
 )
 
@@ -233,7 +233,7 @@ func CreateAnfCapacityPool(ctx context.Context, location, resourceGroupName, acc
 }
 
 // CreateAnfVolume creates an ANF volume within a Capacity Pool
-func CreateAnfVolume(ctx context.Context, location, resourceGroupName, accountName, poolName, volumeName, serviceLevel, subnetID, snapshotID string, protocolTypes []string, volumeUsageQuota int64, unixReadOnly, unixReadWrite bool, tags map[string]*string) (netapp.Volume, error) {
+func CreateAnfVolume(ctx context.Context, location, resourceGroupName, accountName, poolName, volumeName, serviceLevel, subnetID, snapshotID string, protocolTypes []string, volumeUsageQuota int64, unixReadOnly, unixReadWrite bool, tags map[string]*string, dataProtectionObject netapp.VolumePropertiesDataProtection) (netapp.Volume, error) {
 
 	if len(protocolTypes) > 1 {
 		return netapp.Volume{}, fmt.Errorf("only one protocol type is supported at this time")
@@ -280,6 +280,7 @@ func CreateAnfVolume(ctx context.Context, location, resourceGroupName, accountNa
 		SubnetID:       to.StringPtr(subnetID),
 		UsageThreshold: to.Int64Ptr(volumeUsageQuota),
 		CreationToken:  to.StringPtr(volumeName),
+		DataProtection: &dataProtectionObject,
 	}
 
 	future, err := volumeClient.CreateOrUpdate(
@@ -308,11 +309,11 @@ func CreateAnfVolume(ctx context.Context, location, resourceGroupName, accountNa
 }
 
 // UpdateAnfVolume update an ANF volume
-func UpdateAnfVolume(ctx context.Context, location, resourceGroupName, accountName, poolName, volumeName string, volumePropertiesPatch netapp.VolumePatchProperties, tags map[string]*string) (netapp.Volume, error) {
+func UpdateAnfVolume(ctx context.Context, location, resourceGroupName, accountName, poolName, volumeName string, volumePropertiesPatch netapp.VolumePatchProperties, tags map[string]*string) (netapp.VolumesUpdateFuture, error) {
 
 	volumeClient, err := getVolumesClient()
 	if err != nil {
-		return netapp.Volume{}, err
+		return netapp.VolumesUpdateFuture{}, err
 	}
 
 	volume, err := volumeClient.Update(
@@ -329,10 +330,41 @@ func UpdateAnfVolume(ctx context.Context, location, resourceGroupName, accountNa
 	)
 
 	if err != nil {
-		return netapp.Volume{}, fmt.Errorf("cannot update volume: %v", err)
+		return netapp.VolumesUpdateFuture{}, fmt.Errorf("cannot update volume: %v", err)
 	}
 
 	return volume, nil
+}
+
+// AuthorizeReplication - authorizes volume replication
+func AuthorizeReplication(ctx context.Context, resourceGroupName, accountName, poolName, volumeName, remoteVolumeResourceID string) error {
+
+	volumeClient, err := getVolumesClient()
+	if err != nil {
+		return err
+	}
+
+	future, err := volumeClient.AuthorizeReplication(
+		ctx,
+		resourceGroupName,
+		accountName,
+		poolName,
+		volumeName,
+		netapp.AuthorizeRequest{
+			RemoteVolumeResourceID: to.StringPtr(remoteVolumeResourceID),
+		},
+	)
+
+	if err != nil {
+		return fmt.Errorf("cannot authorize volume replication: %v", err)
+	}
+
+	err = future.WaitForCompletionRef(ctx, volumeClient.Client)
+	if err != nil {
+		return fmt.Errorf("cannot get authorize volume replication future response: %v", err)
+	}
+
+	return nil
 }
 
 // CreateAnfSnapshot creates a Snapshot from an ANF volume
@@ -347,7 +379,6 @@ func CreateAnfSnapshot(ctx context.Context, location, resourceGroupName, account
 		ctx,
 		netapp.Snapshot{
 			Location: to.StringPtr(location),
-			Tags:     tags,
 		},
 		resourceGroupName,
 		accountName,
@@ -530,4 +561,56 @@ func WaitForNoANFResource(ctx context.Context, resourceID string, intervalInSec 
 	}
 
 	return fmt.Errorf("exceeded number of retries: %v", retries)
+}
+
+// WaitForANFResource waits for a specified resource to be fully ready following a creation operation.
+func WaitForANFResource(ctx context.Context, resourceID string, intervalInSec int, retries int) error {
+
+	var err error
+
+	for i := 0; i < retries; i++ {
+		time.Sleep(time.Duration(intervalInSec) * time.Second)
+		if uri.IsAnfSnapshot(resourceID) {
+			client, _ := getSnapshotsClient()
+			_, err = client.Get(
+				ctx,
+				uri.GetResourceGroup(resourceID),
+				uri.GetAnfAccount(resourceID),
+				uri.GetAnfCapacityPool(resourceID),
+				uri.GetAnfVolume(resourceID),
+				uri.GetAnfSnapshot(resourceID),
+			)
+		} else if uri.IsAnfVolume(resourceID) {
+			client, _ := getVolumesClient()
+			_, err = client.Get(
+				ctx,
+				uri.GetResourceGroup(resourceID),
+				uri.GetAnfAccount(resourceID),
+				uri.GetAnfCapacityPool(resourceID),
+				uri.GetAnfVolume(resourceID),
+			)
+		} else if uri.IsAnfCapacityPool(resourceID) {
+			client, _ := getPoolsClient()
+			_, err = client.Get(
+				ctx,
+				uri.GetResourceGroup(resourceID),
+				uri.GetAnfAccount(resourceID),
+				uri.GetAnfCapacityPool(resourceID),
+			)
+		} else if uri.IsAnfAccount(resourceID) {
+			client, _ := getAccountsClient()
+			_, err = client.Get(
+				ctx,
+				uri.GetResourceGroup(resourceID),
+				uri.GetAnfAccount(resourceID),
+			)
+		}
+
+		// In this case, we exit when there is no error
+		if err == nil {
+			return nil
+		}
+	}
+
+	return fmt.Errorf("resource still not found after number of retries: %v, error: %v", retries, err)
 }