+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 ## Azure NetAppFiles SDK Sample for Go Changelog
 
+- [1.0.1 (2020-10-12)](#101-2020-10-12)
 - [1.0.0 (2020-02-07)](#100-2020-02-07)
+
+# 1.0.1 (2020-10-12)
+
+*Features*
+* API Version bump up
+* sdkutils.go updated to be in sync with other ANF Go samples
+
+*Bug Fixes*
+* GetAnfCapacityPool bug fix when it sending the wrong string (account)
+* Changes related to new arguments on sdkutils.go
+
+*Breaking Changes*
+* N/A
 
 # 1.0.0 (2020-02-07)